Alteração na estrutura do projeto
