+++ src/converters/time/pages/time-converter-page/time-converter-page.component.ts
@@ -1,7 +1,7 @@
 import { CommonModule, NgFor } from '@angular/common';
 import { Component, OnInit } from '@angular/core';
 import { FormsModule } from '@angular/forms';
-import { PageBase } from '../pageBase';
+import { PageBase } from '../../shared/pages/pageBase';
 
 @Component({
   selector: 'time-converter-page',
