+++ src/generators/cpf/pages/cpf-generator-page/cpf-generator-page.component.scss
