+++ src/shared/components/cookie-consent/cookie-consent.component.scss
