+++ src/shared/components/main-menu/main-menu.component.scss
