+++ src/shared/pages/home-page/home-page.component.scss
