+++ src/shared/pages/main-layout/main-layout.component.scss
