+++ src/shared/services/menu/menu.service.ts
@@ -2,7 +2,7 @@
 import { Injectable } from '@angular/core';
 import { VolumeConverterService } from 'src/converters/volume/services/volume-converter/volume-converter.service';
 import { WeightMassConverterService } from 'src/converters/weight-and-mass/services/weight-mass/weight-mass-converter.service';
-import { MenuConversor } from 'src/layout/models/menuConversor';
+import { MenuConversor } from 'src/shared/models/menuConversor';
 
 @Injectable({
   providedIn: 'root'
