+++ src/validators/cnpj/pages/cnpj-validator-page/cnpj-validator-page.component.ts
@@ -1,6 +1,6 @@
 import { Component, OnInit } from '@angular/core';
 import { FormsModule } from '@angular/forms';
-import { PageBase } from 'src/pages/pageBase';
+import { PageBase } from 'src/shared/pages/pageBase';
 
 @Component({
   selector: 'cnpj-validator-page',
