+++ src/validators/cpf/pages/cpf-validator-page/cpf-validator-page.component.ts
@@ -1,7 +1,7 @@
 import { NgIf } from '@angular/common';
 import { Component, OnInit } from '@angular/core';
 import { FormsModule } from '@angular/forms';
-import { PageBase } from 'src/pages/pageBase';
+import { PageBase } from 'src/shared/pages/pageBase';
 
 @Component({
   selector: 'cpf-validator-page',
